+++ layers/GatoGraphQLForWP/phpunit-packages/gatographql/tests/Integration/AllowedTaxonomyMetaModifyPluginSettingsFixtureEndpointWebserverRequestTest.php
@@ -12,7 +12,7 @@ class AllowedTaxonomyMetaModifyPluginSettingsFixtureEndpointWebserverRequestTest
 
     protected static function getEndpoint(): string
     {
-        return 'graphql/';
+        return 'graphql';
     }
 
     protected function getSettingsKey(): string