+++ layers/GatoGraphQLForWP/phpunit-packages/gatographql/tests/Integration/DisableSchemaModulesOnPrivateEndpointTestOnAdminEndpointsFixtureEndpointWebserverRequestTest.php
@@ -6,13 +6,10 @@ namespace PHPUnitForGatoGraphQL\GatoGraphQL\Integration;
 
 class DisableSchemaModulesOnPrivateEndpointTestOnAdminEndpointsFixtureEndpointWebserverRequestTest extends AbstractDisableSchemaModulesOnPrivateEndpointsFixtureEndpointWebserverRequestTestCase
 {
+    use DisableSchemaModulesOnPrivateEndpointHasChangeFixtureEndpointWebserverRequestTestTrait;
+
     protected static function getEndpoint(): string
     {
         return 'wp-admin/edit.php?page=gatographql&action=execute_query';
-    }
-
-    protected static function getResponseFixtureFolder(): string
-    {
-        return __DIR__ . '/fixture-disable-schema-modules-on-private-endpoints-has-change';
     }
 }