+++ layers/GatoGraphQLForWP/phpunit-packages/gatographql/tests/Integration/DisabledClientWebserverRequestTest.php
@@ -60,14 +60,6 @@ class DisabledClientWebserverRequestTest extends AbstractDisabledClientWebserver
                 'schema/',
                 404,
             ],
-            'custom-endpoint-graphiql' => [
-                'graphql/customers/penguin-books/?view=graphiql',
-                200,
-            ],
-            'custom-endpoint-voyager' => [
-                'graphql/customers/penguin-books/?view=schema',
-                200,
-            ],
         ];
     }