+++ layers/GatoGraphQLForWP/phpunit-packages/gatographql/tests/Integration/SettingsModifyPluginSettingsFixtureEndpointWebserverRequestTest.php
@@ -11,7 +11,7 @@ class SettingsModifyPluginSettingsFixtureEndpointWebserverRequestTest extends Ab
 {
     protected static function getEndpoint(): string
     {
-        return 'graphql/';
+        return 'graphql';
     }
 
     protected static function getFixtureFolder(): string