+++ layers/GatoGraphQLForWP/phpunit-packages/gatographql/tests/Integration/WordPressAuthenticatedUserEndpoints.php
@@ -9,6 +9,5 @@ class WordPressAuthenticatedUserEndpoints
     public const ENDPOINTS = [
         'admin-client' => 'wp-admin/edit.php?page=gatographql&action=execute_query',
         'admin-unrestricted-client' => 'wp-admin/edit.php?page=gatographql&action=execute_query&endpoint_group=pluginOwnUse',
-        'editing-persisted-query-client' => 'wp-admin/edit.php?page=gatographql&action=execute_query&endpoint_group=persistedQuery&persisted_query_id=65',
     ];
 }