+++ layers/GatoGraphQLForWP/plugins/gatographql/editor-scripts/persisted-query-editor-components/docs/en/index.js
@@ -1,2 +1,0 @@
-export * from './welcome-guide.md' ;
-export * from './schema-config-options.md' ;