+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Registries/CustomEndpointAnnotatorRegistry.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GatoGraphQL\GatoGraphQL\Registries;
-
-class CustomEndpointAnnotatorRegistry extends AbstractEndpointAnnotatorRegistry implements CustomEndpointAnnotatorRegistryInterface
-{
-}