+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Registries/PersistedQueryEndpointBlockRegistryInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GatoGraphQL\GatoGraphQL\Registries;
-
-interface PersistedQueryEndpointBlockRegistryInterface extends BlockRegistryInterface
-{
-}