+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Services/Blocks/PersistedQueryEndpointEditorBlockServiceTagInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace GatoGraphQL\GatoGraphQL\Services\Blocks;
 
+use GatoGraphQL\GatoGraphQL\Services\Blocks\EditorBlockInterface;
+
 interface PersistedQueryEndpointEditorBlockServiceTagInterface extends EditorBlockInterface
 {
 }